+++ naga/src/front/wgsl/lower/mod.rs
@@ -1022,7 +1022,7 @@ enum LoweredGlobalDecl {
     Const(Handle<ir::Constant>),
     Override(Handle<ir::Override>),
     Type(Handle<ir::Type>),
-    EntryPoint,
+    EntryPoint(usize),
 }
 
 enum Texture {
@@ -1130,6 +1130,10 @@ impl<'source, 'temp> Lowerer<'source, 'temp> {
             layouter: &mut proc::Layouter::default(),
             global_expression_kind_tracker: &mut proc::ExpressionKindTracker::new(),
         };
+        if !tu.doc_comments.is_empty() {
+            ctx.module.get_or_insert_default_doc_comments().module =
+                tu.doc_comments.iter().map(|s| s.to_string()).collect();
+        }
 
         for decl_handle in self.index.visit_ordered() {
             let span = tu.decls.get_span(decl_handle);
@@ -1138,6 +1142,29 @@ impl<'source, 'temp> Lowerer<'source, 'temp> {
             match decl.kind {
                 ast::GlobalDeclKind::Fn(ref f) => {
                     let lowered_decl = self.function(f, span, &mut ctx)?;
+                    if !f.doc_comments.is_empty() {
+                        match lowered_decl {
+                            LoweredGlobalDecl::Function { handle, .. } => {
+                                ctx.module
+                                    .get_or_insert_default_doc_comments()
+                                    .functions
+                                    .insert(
+                                        handle,
+                                        f.doc_comments.iter().map(|s| s.to_string()).collect(),
+                                    );
+                            }
+                            LoweredGlobalDecl::EntryPoint(index) => {
+                                ctx.module
+                                    .get_or_insert_default_doc_comments()
+                                    .entry_points
+                                    .insert(
+                                        index,
+                                        f.doc_comments.iter().map(|s| s.to_string()).collect(),
+                                    );
+                            }
+                            _ => {}
+                        }
+                    }
                     ctx.globals.insert(f.name.name, lowered_decl);
                 }
                 ast::GlobalDeclKind::Var(ref v) => {
@@ -1173,6 +1200,15 @@ impl<'source, 'temp> Lowerer<'source, 'temp> {
                         span,
                     );
 
+                    if !v.doc_comments.is_empty() {
+                        ctx.module
+                            .get_or_insert_default_doc_comments()
+                            .global_variables
+                            .insert(
+                                handle,
+                                v.doc_comments.iter().map(|s| s.to_string()).collect(),
+                            );
+                    }
                     ctx.globals
                         .insert(v.name.name, LoweredGlobalDecl::Var(handle));
                 }
@@ -1203,6 +1239,15 @@ impl<'source, 'temp> Lowerer<'source, 'temp> {
 
                     ctx.globals
                         .insert(c.name.name, LoweredGlobalDecl::Const(handle));
+                    if !c.doc_comments.is_empty() {
+                        ctx.module
+                            .get_or_insert_default_doc_comments()
+                            .constants
+                            .insert(
+                                handle,
+                                c.doc_comments.iter().map(|s| s.to_string()).collect(),
+                            );
+                    }
                 }
                 ast::GlobalDeclKind::Override(ref o) => {
                     let explicit_ty =
@@ -1249,6 +1294,15 @@ impl<'source, 'temp> Lowerer<'source, 'temp> {
                     let handle = self.r#struct(s, span, &mut ctx)?;
                     ctx.globals
                         .insert(s.name.name, LoweredGlobalDecl::Type(handle));
+                    if !s.doc_comments.is_empty() {
+                        ctx.module
+                            .get_or_insert_default_doc_comments()
+                            .types
+                            .insert(
+                                handle,
+                                s.doc_comments.iter().map(|s| s.to_string()).collect(),
+                            );
+                    }
                 }
                 ast::GlobalDeclKind::Type(ref alias) => {
                     let ty = self.resolve_named_ast_type(
@@ -1469,7 +1523,9 @@ impl<'source, 'temp> Lowerer<'source, 'temp> {
                 workgroup_size_overrides,
                 function,
             });
-            Ok(LoweredGlobalDecl::EntryPoint)
+            Ok(LoweredGlobalDecl::EntryPoint(
+                ctx.module.entry_points.len() - 1,
+            ))
         } else {
             let handle = ctx.module.functions.append(function, span);
             Ok(LoweredGlobalDecl::Function {
@@ -2086,7 +2142,7 @@ impl<'source, 'temp> Lowerer<'source, 'temp> {
                     }
                     LoweredGlobalDecl::Function { .. }
                     | LoweredGlobalDecl::Type(_)
-                    | LoweredGlobalDecl::EntryPoint => {
+                    | LoweredGlobalDecl::EntryPoint(_) => {
                         return Err(Box::new(Error::Unexpected(span, ExpectedToken::Variable)));
                     }
                 };
@@ -2373,7 +2429,7 @@ impl<'source, 'temp> Lowerer<'source, 'temp> {
                 function_span,
                 ExpectedToken::Function,
             ))),
-            Some(&LoweredGlobalDecl::EntryPoint) => {
+            Some(&LoweredGlobalDecl::EntryPoint(_)) => {
                 Err(Box::new(Error::CalledEntryPoint(function_span)))
             }
             Some(&LoweredGlobalDecl::Function {
@@ -3581,6 +3637,8 @@ impl<'source, 'temp> Lowerer<'source, 'temp> {
         let mut struct_alignment = proc::Alignment::ONE;
         let mut members = Vec::with_capacity(s.members.len());
 
+        let mut doc_comments: Vec<Option<Vec<String>>> = Vec::new();
+
         for member in s.members.iter() {
             let ty = self.resolve_ast_type(member.ty, &mut ctx.as_const())?;
 
@@ -3623,6 +3681,11 @@ impl<'source, 'temp> Lowerer<'source, 'temp> {
             offset = member_alignment.round_up(offset);
             struct_alignment = struct_alignment.max(member_alignment);
 
+            if !member.doc_comments.is_empty() {
+                doc_comments.push(Some(
+                    member.doc_comments.iter().map(|s| s.to_string()).collect(),
+                ));
+            }
             members.push(ir::StructMember {
                 name: Some(member.name.name.to_owned()),
                 ty,
@@ -3646,6 +3709,14 @@ impl<'source, 'temp> Lowerer<'source, 'temp> {
             },
             span,
         );
+        for (i, c) in doc_comments.drain(..).enumerate() {
+            if let Some(comment) = c {
+                ctx.module
+                    .get_or_insert_default_doc_comments()
+                    .struct_members
+                    .insert((handle, i), comment);
+            }
+        }
         Ok(handle)
     }